Consolidate noise conditioning into NoiseConditionedModel in stochastic_sfno
Address review feedback:
- Rename NoiseConditionedSFNO to NoiseConditionedModel with an optional
  inverse_sht parameter (None = gaussian, provided = isotropic noise).
  Keep NoiseConditionedSFNO as a backward-compatible alias.
- Delete noise_conditioned.py; local_net.py imports from stochastic_sfno.
- Remove identity grid_mapping dict in LocalNet builder.
- Use dist.get_isht() for creating the inverse SHT in LocalNet builder.
- Delete make_noise_generator helper to avoid type-unsafe ValueError.
